# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: InstaVEG - Hyperlocal grocery marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (seller/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # PAYMENT
  /orders/{id}/payment/create:
    post:
      operationId: createPayment
      tags: [orders]
      summary: Create Razorpay payment order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Razorpay order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrder"
  /orders/{id}/payment/verify:
    post:
      operationId: verifyPayment
      tags: [orders]
      summary: Verify Razorpay payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerification"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # WISHLIST
...
components:
  schemas:
...
    Order:
      type: object
      required: [id, userId, status, paymentStatus, total, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        sellerId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, confirmed, preparing, dispatched, delivered, cancelled]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        paymentMethod:
          type: ["string", "null"]
        razorpayOrderId:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        addressId:
          type: ["integer", "null"]
        addressSnapshot:
          type: ["string", "null"]
        subtotal:
          type: number
        deliveryFee:
          type: number
        platformFee:
          type: number
        total:
          type: number
        eta:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        sellerName:
          type: ["string", "null"]
        buyerName:
          type: ["string", "null"]
    OrderItem:
      type: object
      required: [id, productId, quantity, price]
      properties:
        id:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        price:
          type: number
        productName:
          type: ["string", "null"]
        productImage:
          type: ["string", "null"]
        productUnit:
          type: ["string", "null"]
    OrderInput:
      type: object
      required: [addressId, paymentMethod]
      properties:
        addressId:
          type: integer
        paymentMethod:
          type: string
          enum: [razorpay, cod]
        notes:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, dispatched, delivered, cancelled]
        eta:
          type: string

    # PAYMENT
    PaymentOrder:
      type: object
      required: [razorpayOrderId, amount, currency, keyId]
      properties:
        razorpayOrderId:
          type: string
        amount:
          type: integer
        currency:
          type: string
        keyId:
          type: string
    PaymentVerification:
      type: object
      required: [razorpayOrderId, razorpayPaymentId, razorpaySignature]
      properties:
        razorpayOrderId:
          type: string
        razorpayPaymentId:
          type: string
        razorpaySignature:
          type: string

    # REVIEW
...